pyproject.toml
[tool.poetry]
name = "Livraria Amazonia"
version = "0.1.0"
description = ""
authors = ["Caio Schmitz <caioschmitz09@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
python-magic = "^0.4.27"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
